requirements.txt
PySide6
pyautogui
sounddevice
numpy
python-vlc
pyexcel-ods3
soundfile
